Fixed compilation warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53970 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/test/test_ncbistr.cpp
@@ -2922,7 +2922,7 @@ BOOST_AUTO_TEST_CASE(s_StringToDoubleSpeed)
     };
     double ssr_min[sizeof(ssr)/sizeof(ssr[0])];
     double ssr_max[sizeof(ssr)/sizeof(ssr[0])];
-    const int TESTS = ArraySize(ss);
+    const size_t TESTS = ArraySize(ss);
 
     int flags = NStr::fConvErr_NoThrow|NStr::fAllowLeadingSpaces;
     double v;
